test: show missing stream event On streams it is possible to receive ERR|HUP after EOF. For named pipes this is the remote end calling shutdown() then close(). For TCP this is receiving a FIN (half closed) and then a RST (full closed).
Attempts to write to these streams would generate an EPIPE, but if just
passively waiting for events (e.g. server off waiting for the result of some
long running task) it could be missed. In the case of the server running a long
task, it would be desireable to know about these events so that long running
work can be stopped.

This commit demonstrates the problem with tests for TCP and named pipes. The
next commit will introduce a fix for the issue to allow events to be raised.