stream: catch hup/err after rx EOF
The previous commit introduced a set of tests showing that HUP|ERROR events can
be missed in some cases for streams. This happens when the remote end performs a
half close via shutdown(..., SHUT_WR, ...) followed by a close() for named pipes
or a SO_LINGER enabled close() for TCP which triggers a RST. It can happen in
some cases for TCP with middleware (firewalls, proxies) as well.

This commit introduces a workaround to allow those events to be delivered via
the read callback. Semantically this makes more sense as a new event type but
adding one introduces potential ABI incompatibility. The semantics of "error
on read callback" already exist and are expected by many 3rd party integrations.

The specific conditionts for triggering this case and reasons it is impotant
for some systems to be able to catch are:
The first call to shutdown(SHUT_WR) puts the stream in a half closed state.
For the receiver (libuv in this case) they can no longer read but can write.
This is rightfully a UV_EOF state and writes will still succeed.

The follow-on close()/RST can be detected once a uv_write() is attempted, but
not all systems will have anything to send back when they are interseted in
these cases. The simplest example is a server running a long resource intensive
background job which can be stopped once the result can no longer be returned.
Prior to this change, the server would not find out the result is unreturnable
until after it has completed the work and attempted the write. After this change
the server knows immeidately that the work can be stopped and resources freed.